core(lantern): move LanternMetric in lib/lantern
